현지 : 카테고리 버튼 추가 JS

+++ src/com/petmet/web/controller/admin/community/category/ListController.java
@@ -1,6 +1,7 @@
 package com.petmet.web.controller.admin.community.category;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.ServletException;
@@ -9,6 +10,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.petmet.web.entity.BoardCategory;
 import com.petmet.web.entity.BoardCategoryView;
 import com.petmet.web.service.BoardCategoryService;
 
@@ -26,15 +28,39 @@ public class ListController extends HttpServlet{
 	
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String[] dels = request.getParameterValues("del");
-		int[] ids = new int[dels.length];
-		
-		for(int i=0; i < dels.length; i++)
-			ids[i] = Integer.parseInt(dels[i]);
-		
 		BoardCategoryService service = new BoardCategoryService();
-		service.deleteList(ids);
+
+		String button = request.getParameter("button");
 		
+		String[] dels = request.getParameterValues("del");
+		String[] newNames = request.getParameterValues("new-name");
+		String[] names = request.getParameterValues("name");
+		
+		// --------------------- 삭제 ---------------------
+		if (dels != null) {
+			int[] ids = new int[dels.length];
+
+			for (int i = 0; i < dels.length; i++)
+				ids[i] = Integer.parseInt(dels[i]);
+
+			service.deleteList(ids);
+		}
+		
+		// --------------------- 삽입 ---------------------
+		if(newNames != null) {
+			List<BoardCategory> list = new ArrayList<BoardCategory>();
+			
+			for (int i = 0; i < newNames.length; i++) {
+				BoardCategory b = new BoardCategory(i, newNames[i]);
+				list.add(b);
+			}
+			
+			service.insertList(list);
+		}
+		
+		// --------------------- 수정 ---------------------
+		
+		// --------------------- 요청 ---------------------
 		response.sendRedirect("list");
 	}
 }
